requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
PyMuPDF==1.24.5
openai==1.35.3
gspread==6.1.2
google-auth==2.30.0
python-dotenv==1.0.1
